scripts/blog_server/pyproject.toml
[project]
name = "blog-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli,server]>=1.9.0",
    "python-frontmatter>=1.1.0",
    "pyyaml>=6.0.2",
]
